Refactor serialization storage for various DOM types
- Updated the `serialized_storage` method in multiple DOM-related files to remove the unnecessary lifetime parameter in `StructuredData`.
- Adjusted the handling of `create_buffer_source` to use the correct types in `dommatrixreadonly.rs`, `dompoint.rs`, `domquad.rs`, `domrect.rs`, and `domrectreadonly.rs`.
- Modified the `EventTarget` implementation to simplify the scope chain creation.
- Enhanced the `GlobalScope` to properly handle cross-realm transformations.
- Fixed the `HTMLScriptElement` to correctly handle module records.
- Updated `MessagePort`, `Permissions`, and `QuotaExceededError` to streamline serialization processes.
- Refined `ReadableByteStreamController` and `ReadableStream` to utilize `Uint8Array` instead of `ArrayBufferViewU8`.
- Improved error handling in `fetch.rs` and `script_module.rs` by ensuring proper exception stack behavior.
- Adjusted the `Window` and `WindowProxy` implementations to enhance proxy traps and improve type handling.
- Updated dependencies in `Cargo.toml` for `egui-file-dialog` to ensure compatibility.
- Refactored file dialog handling in `dialog.rs` to utilize static string references for filter patterns.

+++ components/script/dom/abortsignal.rs
@@ -463,7 +463,7 @@ impl AbortSignalMethods<crate::DomTypeHolder> for AbortSignal {
                 JS_SetPendingException(
                     cx,
                     HandleValue::from_raw(self.abort_reason.handle()),
-                    ExceptionStackBehavior::Capture,
+                    ExceptionStackBehavior::Capture as u32,
                 )
             };
             return Err(Error::JSFailed);
